Downgrade JRE in base image
It turns out that the default version of JRE 11 is not compatible with
even the latest Spark.  As a result, it is downgraded to JRE 8 by using
the Sid repository.

In addition, the images are attempted to be made more
development-friendly by including vim, tmux, and keeping the apt cache
lists.