Split the base image
In the base image, in addition to core developmental dependencies like
the latest GCC or cmake, we also have drudge dependencies like the
latest Python and Spark.

Since the later part is not actually used in the development of core
libraries like libcanon or libparenth, in this commit, the base image is
split into two, with `base` having the same content, and the new
`basecore` added for the development of the core C++ libraries.

For the core base, in this commit, gdb and the pretty-printer from GCC
are added for the convenience of development.